fix $post bug, added "page_for_posts"
In the original filterMenu code, the global $post ended up being set to the post the menu was retrieved from. This caused havoc/unreliable glitches for the rest of the theme/page/post.
Fixed by using a local variable to hold the current post.

Also added support for setting the menus on the "page_for_posts" page, as the original code ended up using the menu defined for the newest blogpost instead. ($post is not set to the page, but the latest blog post on this special page)